Fix object color recognition
This commit introduce change in interpreting single point in WallMap and Door.
Before this commit two points was treated as a wall and single point was not
rendered. Since now single point is treated as wall. This change eliminates
ambiguous color if two walls with different color were placed together.